% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honnorat.R
\name{combat_harmonization}
\alias{combat_harmonization}
\title{Harmonize Vector Images Using ComBat}
\usage{
combat_harmonization(super_sample)
}
\arguments{
\item{super_sample}{A \code{CSuperSample} object containing samples to harmonize.}
}
\value{
A new \code{CSuperSample} object with harmonized vector images.
}
\description{
Applies the ComBat batch effect correction to vector images
in a \code{CSuperSample} object and reconstructs harmonized samples.
}
