% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honnorat.R
\name{format_matr}
\alias{format_matr}
\title{Format a Matrix as a Packed dpoMatrix}
\usage{
format_matr(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A packed \code{dpoMatrix} object.
}
\description{
Converts a matrix to a packed symmetric positive definite matrix
(\code{dpoMatrix}) using the Matrix package.
}
