% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_anova.R
\name{frechet_anova}
\alias{frechet_anova}
\title{Frechet ANOVA Test Statistic}
\usage{
frechet_anova(super_sample)
}
\arguments{
\item{super_sample}{An object of class CSuperSample (see riemtan package)}
}
\value{
A list with two elements:
\describe{
\item{statistic}{The Frechet ANOVA test statistic}
\item{p_value}{The p-value from the chi-squared test}
}
}
\description{
Computes the Frechet ANOVA test statistic and its p-value for a given super-sample object.
}
