% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riem_anova.R
\name{log_wilks_lambda}
\alias{log_wilks_lambda}
\title{Compute the Log Wilks' Lambda Statistic}
\usage{
log_wilks_lambda(super_sample)
}
\arguments{
\item{super_sample}{An object of class CSuperSample}
}
\value{
A numeric value representing the log Wilks' lambda statistic.
}
\description{
Calculates the log of Wilks' lambda statistic
for a given \code{super_sample} object.
This function ensures that the within-group
and total covariance matrices are computed,
then computes the difference of their log determinants.
}
\details{
Wilks' lambda is a test statistic for the ANOVA test decribed in (to appear)
}
