% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riem_anova.R
\name{one_bootstrap}
\alias{one_bootstrap}
\title{Bootstrap Statistic for a Super Sample}
\usage{
one_bootstrap(x, hat_sigma, hat_gamma, geom, stat_fun)
}
\arguments{
\item{x}{An object of class \code{CSuperSample},
representing the original super sample.}

\item{hat_sigma}{The mean parameter for the normal distribution
used in resampling.}

\item{hat_gamma}{The covariance parameter for the normal distribution
used in resampling.}

\item{geom}{The geometry parameter to be passed to \code{riemtan::rspdnorm}.}

\item{stat_fun}{A function to compute a statistic
on the resulting \code{CSuperSample} object.}
}
\value{
The value returned by \code{stat_fun} when applied
to the bootstrapped super sample.
}
\description{
For each subsample in a given super sample,
draws a normal subsample of the same size
using the provided parameters, collects them into a new super sample,
and computes a specified statistic on the resulting super sample.
}
\details{
This function performs a parametric bootstrap by resampling
each subsample in \code{x} according to the specified parameters,
then aggregates the resampled data into a new
\code{CSuperSample} and computes the desired statistic.
}
