% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riem_anova.R
\name{pillais_trace}
\alias{pillais_trace}
\title{Compute Pillai's Trace Statistic}
\usage{
pillais_trace(super_sample)
}
\arguments{
\item{super_sample}{An object of class CSuperSample}
}
\value{
A numeric value representing Pillai's trace statistic.
}
\description{
Calculates Pillai's trace statistic for a given \code{super_sample} object.
This function ensures that the within-group
and total covariance matrices are computed,
then computes the sum of the eigenvalues of
the matrix (Total - Within) \%*\% solve(Total).
}
\details{
Pillai's trace is a test statistic for
the ANOVA test decribed in (to appear).
}
