% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honnorat.R
\name{ts2corr}
\alias{ts2corr}
\title{Compute OAS-Shrunk Correlation Matrix from Time Series}
\usage{
ts2corr(ts)
}
\arguments{
\item{ts}{A numeric matrix representing time series data.}
}
\value{
A shrunk correlation matrix.
}
\description{
Normalizes a time series matrix and computes
its OAS-shrunk correlation matrix.
}
