% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riem_anova.R
\name{riem_anova}
\alias{riem_anova}
\title{Compute p-values using bootstrap}
\usage{
riem_anova(ss, stat_fun = log_wilks_lambda, den = 100)
}
\arguments{
\item{ss}{An object of class \code{CSuperSample}.}

\item{stat_fun}{A function to compute a statistic
on the \code{CSuperSample} object (default: \code{log_wilks_lambda}).}

\item{den}{The number of bootstrap samples to generate
for estimating the p-value.}
}
\value{
numeric A bootstrap p-value.
}
\description{
Computes a bootstrap-based p-value and for a given super sample.
The statistic used for the bootstrap can be specified
via the \code{stat_fun} argument.
}
\details{
The function computes the statistic on the observed data
and compares it to the distribution
of statistics computed on bootstrapped samples.
}
